Reorganize resources.
