+++ .eslintrc.js
@@ -1,15 +1,17 @@
 // eslint-disable-next-line no-undef
 module.exports = {
   root: true,
-  parser: '@typescript-eslint/parser',
-  plugins: [
-    '@typescript-eslint',
-  ],
-  extends: [
-    'eslint:recommended',
-    'plugin:@typescript-eslint/recommended',
-  ],
+  parser: "@typescript-eslint/parser",
+  plugins: ["@typescript-eslint"],
+  extends: ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
   rules: {
-    "@typescript-eslint/no-unused-vars": ["warn", { "varsIgnorePattern": "^_" } ]
-  }
+    "@typescript-eslint/no-unused-vars": ["warn", { varsIgnorePattern: "^_" }],
+    "sort-imports": [
+      "error",
+      {
+        ignoreCase: true,
+        ignoreDeclarationSort: true,
+      },
+    ],
+  },
 };
