+++ src/resources/2012/RainDoh.ts
@@ -1,8 +1,8 @@
 import { use } from "kolmafia";
 
-import { getFoldGroup, have as haveItem } from "../lib";
-import { get } from "../property";
-import { $item } from "../template-string";
+import { getFoldGroup, have as haveItem } from "../../lib";
+import { get } from "../../property";
+import { $item } from "../../template-string";
 
 export const box = $item`Rain-Doh black box`;
