+++ src/resources/2014/WinterGarden.ts
@@ -1,7 +1,7 @@
-import { Copier } from "../Copier";
-import { get } from "../property";
-import { have as haveItem, haveInCampground } from "../lib";
-import { $item } from "../template-string";
+import { Copier } from "../../Copier";
+import { get } from "../../property";
+import { haveInCampground, have as haveItem } from "../../lib";
+import { $item } from "../../template-string";
 
 export function have(): boolean {
   return haveInCampground($item`packet of winter seeds`);
