+++ src/resources/index.ts
@@ -1,10 +1,10 @@
-export * as ChateauMantegna from "./ChateauMantegna";
-export * as ObtuseAngel from "./ObtuseAngel";
-export * as RainDoh from "./RainDoh";
-export * as SourceTerminal from "./SourceTerminal";
-export * as SpookyPutty from "./SpookyPutty";
-export * as TunnelOfLove from "./TunnelOfLove";
-export * as WinterGarden from "./WinterGarden";
-export * as Witchess from "./Witchess";
+export * as ChateauMantegna from "./2015/ChateauMantegna";
+export * as ObtuseAngel from "./2011/ObtuseAngel";
+export * as RainDoh from "./2012/RainDoh";
+export * as SourceTerminal from "./2016/SourceTerminal";
+export * as SpookyPutty from "./2009/SpookyPutty";
+export * as TunnelOfLove from "./2017/TunnelOfLove";
+export * as WinterGarden from "./2014/WinterGarden";
+export * as Witchess from "./2016/Witchess";
 
 export * from "./putty-likes";
